requirements.txt
torch
torchvision
pillow
customtkinter
